backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5
whitenoise==6.6.0
pyodbc==5.0.1
python-docx==1.1.0
docxtpl==0.16.7
reportlab==4.0.7
Pillow==10.1.0
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0